% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqdplot.R
\name{ggseqdplot}
\alias{ggseqdplot}
\title{Render sequence distribution plot with ggplot2}
\usage{
ggseqdplot(
  seqdata,
  group = NULL,
  weighted = TRUE,
  with.missing = FALSE,
  border = TRUE
)
}
\arguments{
\item{seqdata}{seqdata State sequence object (class \code{stslist}) created with the \code{TraMineR::seqdef} function.}

\item{group}{Grouping variable of length equal to the number of sequences.
When not NULL, a distinct plot is generated for each level of group.}

\item{weighted}{specifies if weights defined when generating the sequence object should be used for computing the state distributions. Default is \code{TRUE}, i.e. if available weights are used}

\item{with.missing}{specifies if missing states should be considered when computing the state distributions.}

\item{border}{if \code{TRUE} (default) bars are plotted with black outline}
}
\value{
A sequence distribution plot. If stored as object the resulting list
object also contains the data (long format) used for rendering the plot
}
\description{
Function for rendering state distribution plots with \code{{ggplot2}} instead of base
R's \code{plot} function that is used by \code{TraMineR::seqplot}.
}
\details{
The function uses \code{TraMineR::seqstatd} to obtain state distributions. Obviously this requires that the
input data (\code{seqdata}) is stored as state sequence object (class \code{stslist}) created with the \code{TraMineR::seqdef} function.
}
\examples{

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# Examples from TraMineR::seqplot

library(TraMineR)
library(ggplot2)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# Examples from TraMineR::seqplot

# actcal data set
data(actcal)

# We use only a sample of 300 cases
set.seed(1)
actcal <- actcal[sample(nrow(actcal),300),]
actcal.lab <- c("> 37 hours", "19-36 hours", "1-18 hours", "no work")
actcal.seq <- seqdef(actcal,13:24,labels=actcal.lab)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# state distribution plots; grouped by sex
# with TraMineR::seqplot
seqdplot(actcal.seq, group=actcal$sex)
# with ggseqplot
ggseqdplot(actcal.seq, group=actcal$sex)
# with ggseqplot and weights turned off
ggseqdplot(actcal.seq, group=actcal$sex, weighted = FALSE)

# make use of ggplot functions for modifying the plot
ggseqdplot(actcal.seq) +
  scale_x_discrete(labels = month.abb) +
  labs(title = "State distribution plot",
       x = "Month") +
  guides(fill=guide_legend(title="Alphabet")) +
  theme_classic() +
  theme(plot.title = element_text(size = 30,
                                  margin=margin(0,0,20,0)),
        plot.title.position = "plot")

}
