% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqiplot.R
\name{ggseqiplot}
\alias{ggseqiplot}
\title{Render sequence index plot with ggplot2}
\usage{
ggseqiplot(
  seqdata,
  group = NULL,
  sortv = NULL,
  weighted = TRUE,
  border = FALSE,
  facet_scale = "free_y",
  facet_ncol = NULL,
  facet_nrow = NULL
)
}
\arguments{
\item{seqdata}{State sequence object (class \code{stslist}) created with the \code{TraMineR::seqdef} function.}

\item{group}{Grouping variable for rendering faceted plots.
When not NULL, a distinct plot is generated for each level of group.}

\item{sortv}{Vector of numerical values for sorting the sequences}

\item{weighted}{Controls if weights should be used}

\item{border}{if \code{TRUE} bars are plotted with black outline}

\item{facet_scale}{Specifies if y-scale in faceted plot should be free ("free_y" = default) or "fixed"}

\item{facet_ncol}{Number of columns in faceted (i.e. grouped) plot}

\item{facet_nrow}{Number of rows in faceted (i.e. grouped) plot}
}
\value{
A sequence index plot. If stored as object the resulting list object
also contains the data (spell format) used for rendering the plot.
}
\description{
Function for rendering sequence index plots with \code{{ggplot2}} instead of base
R's \code{plot} function that is used by \code{TraMineR::seqplot}.
}
\details{
The function uses \code{TraMineR::seqformat} to reshape \code{seqdata} stored in wide format into
a spell/episode format. If \code{border=TRUE} the data are reshaped into the long format, i.e. for every sequence
each row in the data represents one specific sequence position. For example, if we have 5 sequences of length 10,
the long file will have 50 rows. In the case of sequences of unequal length not every sequence will contribute
the same number of rows to the long data.

The reshaped data are used as input for rendering the index plot using ggplot2's \code{geom_rect}.
\code{ggseqiplot} uses \code{geom_rect} instead of \code{geom_tile} because this allows for a straight forward
implementation of weights. If weights are specified for \code{seqdata} and \code{weighted=TRUE} the sequence height
corresponds to its weight.

Note that the default aspect ratio of \code{ggseqiplot} is different from \code{TraMineR::seqIplot}. This is most
obvious when \code{border=TRUE}. You can change the ratio either by adding code to \code{ggseqiplot}
or by specifying the ratio when saving the code with \code{ggsave}.
}
\examples{

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# Examples from TraMineR::seqplot

library(TraMineR)

# actcal data set
data(actcal)

# We use only a sample of 300 cases
set.seed(1)
actcal <- actcal[sample(nrow(actcal),300),]
actcal.lab <- c("> 37 hours", "19-36 hours", "1-18 hours", "no work")
actcal.seq <- seqdef(actcal,13:24,labels=actcal.lab)

# ex1 using weights
data(ex1)
ex1.seq <- seqdef(ex1, 1:13, weights=ex1$weights)

# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~


# sequences sorted by age in 2000 and grouped by sex
# with TraMineR::seqplot
seqIplot(actcal.seq, group=actcal$sex,sortv=actcal$age00)
# with ggseqplot
ggseqiplot(actcal.seq, group=actcal$sex, sortv=actcal$age00)

# sequences of unequal length with missing state, and weights
seqIplot(ex1.seq)
ggseqiplot(ex1.seq)

# ... turn weights off add border
seqIplot(ex1.seq, weighted = FALSE, border = TRUE)
ggseqiplot(ex1.seq, weighted = FALSE, border = TRUE)

}
\author{
Marcel Raab
}
