% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqplot-package.R
\docType{package}
\name{ggseqplot-package}
\alias{ggseqplot}
\alias{ggseqplot-package}
\title{ggseqplot: Render Sequence Plots using 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of wrapper functions that mainly re-produces most of the sequence plots rendered with TraMineR::seqplot(). Whereas 'TraMineR' uses base R to produce the plots this library draws on 'ggplot2'. The plots are produced on the basis of a sequence object defined with TraMineR::seqdef(). The package automates the reshaping and plotting of sequence data. Resulting plots are of class 'ggplot', i.e. components can be added and tweaked using '+' and regular 'ggplot2' functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://maraab23.github.io/ggseqplot/}
  \item \url{https://github.com/maraab23/ggseqplot}
  \item Report bugs at \url{https://github.com/maraab23/ggseqplot/issues}
}

}
\author{
\strong{Maintainer}: Marcel Raab \email{marcel.raab@gmail.com} (\href{https://orcid.org/0000-0002-3097-1591}{ORCID})

}
\keyword{internal}
