% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggseqplot-package.R, R/ggseqplot.R
\docType{package}
\name{ggseqplot-package}
\alias{ggseqplot}
\alias{ggseqplot-package}
\title{ggseqplot: Render Sequence Plots using ggplot2}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}


This library provides a set of wrapper functions that reproduce some of the sequence plots from \code{\link[TraMineR:seqplot]{TraMineR::seqplot}} and \code{\link[TraMineRextras]{TraMineRextras}} using \code{\link[ggplot2]{ggplot2}}. These plots are produced on the basis of a sequence object defined with \code{\link[TraMineR:seqdef]{TraMineR::seqdef}}. The package automates the reshaping and plotting of sequence data. Resulting plots are of class ggplot, i.e. components can be added and tweaked using `+` and regular ggplot2 functions.

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/maraab23/ggseqplot}
  \item Report bugs at \url{https://github.com/maraab23/ggseqplot/issues}
}

}
\author{
\strong{Maintainer}: Marcel Raab \email{marcel.raab@gmail.com} (\href{https://orcid.org/0000-0002-3097-1591}{ORCID})

}
\keyword{internal}
